+++ back-end/api/auth/login.js
@@ -1,6 +1,6 @@
 const express = require('express');
-const user = require('../../models/user');
-const token = require('../../models/token');
+const user = require('../../config/user');
+const token = require('../../config/token');
 const jwt = require('jsonwebtoken');
 const bcrypt = require('bcrypt');
 const routes = express.Router();
@@ -23,7 +23,7 @@ routes.post('/', (req,res)=>{
                     "username":value.username,
                     "userid":value.userid
                 }
-                const accessToken = jwt.sign(payload,SECRET,{expiresIn:"10s"});
+                const accessToken = jwt.sign(payload,SECRET,{expiresIn:"30m"});
                 const refreshToken = jwt.sign(payload,SECRET,{expiresIn:"1h"});
                 token.refreshTokenSave(userid,refreshToken)
                 .then(console.log)
