+++ back-end/api/posts/posts.js
@@ -1,6 +1,7 @@
 const routes = require('express').Router();
-const post = require('../../models/post')
+const post = require('../../config/post')
 const jwt = require('jsonwebtoken');
+const passport = require('passport');
 const JwtStrategy = require('passport-jwt').Strategy,
     ExtractJwt = require('passport-jwt').ExtractJwt;
 const opts = {}
@@ -14,48 +15,60 @@ opts.secretOrKey = SECRET;
 //FIXME: middleware로 token이 있다면, 글쓰기 가능
 
 /** TODO: 게시물 게시 */
-routes.post('/',(req,res,next)=>{
-    const accessToken = req.headers["authorization"];
-    const [tokenType,tokenValue] = accessToken.split(' ')
-    if(tokenType !=TOKEN_TYPE){
-        res.send({"msg":"login을 해주세요"})
+routes.post('/',passport.authenticate('jwt',{session:false}),(req,res,next)=>{
+    const [tokenType,tokenValue] = req.headers["authorization"].split(' ')
+    const decodeToken = jwt.decode(tokenValue,SECRET);
+    
+    const title = req.body.title;
+    const contents = req.body.contents;
+    if(title == null || contents == null){
+        //빈칸존재
+        res.send({"msg":'err1'})
         return;
     }
-    try{
-        jwt.verify(tokenValue,SECRET);
-        const decodeToken = jwt.decode(tokenValue,SECRET);
-        
-        const title = req.body.title;
-        const content = req.body.content;
-        data = [decodeToken.userid,decodeToken.username,title,content];
-        post.createPost(data)
-        .then((result)=>{
-            res.send({"msg":"good"})
-            return;
-        })
-        .catch()
-    }catch(err){
-        res.send({"msg":"login을 해주세요"})
-    }
+    data = [decodeToken.userid,decodeToken.username,title,contents];
+    
+    post.createPost(data)
+    .then((result)=>{
+        console.log(result)
+        //등록 완료
+        res.send({"msg":"good"})
+        return;
+    })
+    .catch(()=>{
+        res.send({"msg":'err1'})
+    })
 
 })
 /** TODO:  게시물 전체 조회 */
 routes.get('/',(req,res)=>{
-    console.log(req.query.idx)
-    if(req.query.idx == null){
-        post.allPosts()
-        .then((values)=>{
-            res.status(200).send(values)
-        })
-        .catch(()=>{
-            console.log('err')
-        })
+    console.log("asd")
+    if(req.query.idx==null){
+        if(req.query.idx == null){
+            post.allPosts()
+            .then((values)=>{
+                res.status(200).send(values)
+            })
+            .catch(()=>{
+                res.send({"msg":'err1'})
+            })
+        }else{
+            //console.log("YES")
+            post.readPost(req.query.idx)
+            .then((value)=>{
+                res.send(value)
+            })
+        }
     }else{
-        console.log("YES")
+        //쿼리스트링 조회
         post.readPost(req.query.idx)
-        .then((value)=>{
-            res.send(value)
-        })
+            .then((value)=>{
+                console.log(value)
+                res.send(value)
+            })
+            .catch(()=>{
+                res.send({"msg":'err1'})
+            })
     }
 })
 /** TODO: 게시물 수정 */
@@ -66,16 +79,5 @@ routes.put('/',(req,res)=>{
 routes.delete('/',(req,res)=>{
 
 })
-// passport.use(new JwtStrategy(opts, (jwt_payload, done)=> {
-//     console.log(jwt_payload)
-//     user.findUser(jwt_payload.userid)
-//     .then((value)=>{
-//         return done(null,{username:value.username, userid:value.userid})
-//     })
-//     .catch((value)=>{
-//         return done(null,false)
-//     })
-// }));
-
 
 module.exports = routes;
