+++ front-end/src/App.js
@@ -1,23 +1,8 @@
-import logo from './logo.svg';
-import './App.css';
-
+import Home from "./components/Home"
 function App() {
   return (
     <div className="App">
-      <header className="App-header">
-        <img src={logo} className="App-logo" alt="logo" />
-        <p>
-          Edit <code>src/App.js</code> and save to reload.
-        </p>
-        <a
-          className="App-link"
-          href="https://reactjs.org"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Learn React
-        </a>
-      </header>
+        <Home/>
     </div>
   );
 }
