+++ front-end/src/components/Home.js
@@ -1,0 +1,9 @@
+import React, {Component}from "react";
+
+class Home extends Component{
+    render(){
+        return <h1>Home 화면임!</h1>
+    }
+}
+
+export default Home;
