+++ front/board.htm
@@ -1,0 +1,44 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Document</title>
+</head>
+<body>
+    <div class="contents">
+
+    </div>
+    <a href="./protected.htm">back</a>
+    <script>
+        const token = sessionStorage.getItem("token");
+        document.addEventListener('DOMContentLoaded',()=>{
+            sch = location.search
+            var params = new URLSearchParams(sch);
+            const idx = params.get('idx')
+            console.log(idx)
+            console.log(`http://localhost:3000/api/posts?idx=${idx}`)
+            fetch(`http://localhost:3000/api/posts?idx=${idx}`, {
+                method: "GET",
+                headers: {
+                "Content-Type": "application/json",
+                "Authorization": token
+            },
+            })
+            .then((response) => response.json())
+            .then((data) =>{
+                const contents = document.querySelector('.contents')
+                contents.innerHTML=
+                `
+                ${data.idx}번 게시물
+                <h1>title : ${data.title}</h1>
+                <small>date : ${data.date}</small>
+                <h3>username : ${data.username}</h3>
+                <h3>contents : ${data.contents}</h3>
+                `
+            });
+        })
+    </script>
+</body>
+</html>
