+++ front/createboard.htm
@@ -1,0 +1,37 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Document</title>
+</head>
+<body>
+    <input type="text" class="title" placeholder="title"><br>
+    <textarea class="ct" rows="5" cols="30" name="contents"></textarea>
+    <button class = "btn">asd</button>
+    <script>
+        const token = sessionStorage.getItem("token");
+        document.querySelector('.btn').addEventListener('click',()=>{
+            const contents = document.querySelector('.ct').value;
+            const ct = contents.replace("\n","<br>")
+            fetch("http://localhost:3000/api/posts", {
+                method: "POST",
+                headers: {
+                "Content-Type": "application/json",
+                "Authorization": token
+            },body: JSON.stringify({
+                "title":document.querySelector('.title').value,
+                "contents":ct
+            }),
+            })
+            .then((response) => response.json())
+            .then((data) =>{
+                console.log(data);
+                location.href = "./protected.htm";
+                alr
+            });
+        })
+    </script>
+</body>
+</html>
